test: cafs
