+++ store/cafs/src/index.ts
@@ -1,15 +1,11 @@
-import { promises as fs, type Stats } from 'fs'
-import path from 'path'
 import type { FileWriteResult, PackageFileInfo } from '@pnpm/cafs-types'
 import getStream from 'get-stream'
-import renameOverwrite from 'rename-overwrite'
 import ssri from 'ssri'
 import { addFilesFromDir } from './addFilesFromDir'
 import { addFilesFromTarball } from './addFilesFromTarball'
 import {
   checkPkgFilesIntegrity,
   type PackageFilesIndex,
-  verifyFileIntegrity,
 } from './checkPkgFilesIntegrity'
 import { readManifestFromStore } from './readManifestFromStore'
 import {
@@ -19,7 +15,7 @@ import {
   getFilePathByModeInCafs,
   modeIsExecutable,
 } from './getFilePathInCafs'
-import { writeFile } from './writeFile'
+import { writeBufferToCafs } from './writeBufferToCafs'
 
 export type { IntegrityLike } from 'ssri'
 
@@ -81,71 +77,4 @@ async function addBufferToCafs (
     integrity
   )
   return { checkedAt, integrity }
-}
-
-async function writeBufferToCafs (
-  locker: Map<string, Promise<number>>,
-  cafsDir: string,
-  buffer: Buffer,
-  fileDest: string,
-  mode: number | undefined,
-  integrity: ssri.IntegrityLike
-): Promise<number> {
-  fileDest = path.join(cafsDir, fileDest)
-  if (locker.has(fileDest)) {
-    return locker.get(fileDest)!
-  }
-  const p = (async () => {
-    // This part is a bit redundant.
-    // When a file is already used by another package,
-    // we probably have validated its content already.
-    // However, there is no way to find which package index file references
-    // the given file. So we should revalidate the content of the file again.
-    if (await existsSame(fileDest, integrity)) {
-      return Date.now()
-    }
-
-    // This might be too cautious.
-    // The write is atomic, so in case pnpm crashes, no broken file
-    // will be added to the store.
-    // It might be a redundant step though, as we verify the contents of the
-    // files before linking
-    //
-    // If we don't allow --no-verify-store-integrity then we probably can write
-    // to the final file directly.
-    const temp = pathTemp(fileDest)
-    await writeFile(temp, buffer, mode)
-    // Unfortunately, "birth time" (time of file creation) is available not on all filesystems.
-    // We log the creation time ourselves and save it in the package index file.
-    // Having this information allows us to skip content checks for files that were not modified since "birth time".
-    const birthtimeMs = Date.now()
-    await renameOverwrite(temp, fileDest)
-    return birthtimeMs
-  })()
-  locker.set(fileDest, p)
-  return p
-}
-
-function pathTemp (file: string): string {
-  const basename = removeSuffix(path.basename(file))
-  return path.join(path.dirname(file), `${basename}_tmp_${process.pid}`)
-}
-
-function removeSuffix (filePath: string): string {
-  const dashPosition = filePath.indexOf('-')
-  if (dashPosition === -1) return filePath
-  return filePath.substring(0, dashPosition)
-}
-
-async function existsSame (filename: string, integrity: ssri.IntegrityLike) {
-  let existingFile: Stats | undefined
-  try {
-    existingFile = await fs.stat(filename)
-  } catch (err) {
-    return false
-  }
-  return verifyFileIntegrity(filename, {
-    size: existingFile.size,
-    integrity,
-  })
 }
